# docs/openapi/wix-landys-pro.yaml
openapi: 3.1.0
info:
  title: Landy's Pro Wix Estimate Intake
  version: 2.0.0
  description: >
    Server-to-server estimate intake with automatic budget tiers, attachments,
    and contractor sync. See docs/wix-final-integration-contract.md.
servers:
  - url: https://{landysProHost}
    variables:
      landysProHost:
        default: example.invalid
        description: Environment-specific Landy's Pro host supplied by the operator.
paths:
  /api/integrations/wix/estimate-requests:
    post:
      operationId: createWixEstimateRequest
      summary: Persist a Wix estimate request for safe review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WixEstimateRequest"
      responses:
        "202":
          description: Request persisted or identical request replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateAccepted"
        "400":
          description: Invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bearer credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: External request ID reused with a different payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Content-Type is not application/json.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failed or an active taxonomy code was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Wix estimate integration is disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    WixEstimateRequest:
      type: object
      additionalProperties: false
      required:
        - source
        - externalRequestId
        - email
        - propertyZip
        - landTypeCode
        - projectTypeCode
        - budget
        - timeline
        - urgency
        - description
      properties:
        source:
          type: string
          enum:
            - general/get-three-estimates
            - direct-contractor-profile-request
        externalRequestId:
          type: string
          minLength: 1
          maxLength: 160
        firstName:
          type:
            - string
            - "null"
          maxLength: 80
        lastName:
          type:
            - string
            - "null"
          maxLength: 80
        phone:
          type:
            - string
            - "null"
          maxLength: 40
        email:
          type: string
          format: email
          maxLength: 320
        propertyZip:
          type: string
          pattern: '^\d{5}(?:-\d{4})?$'
        contractorCategoryCode:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 80
        landTypeCode:
          type: string
          minLength: 1
          maxLength: 80
        projectTypeCode:
          type: string
          minLength: 1
          maxLength: 80
          description: Live work type code or Wix label (e.g. repair, New Build). Legacy land project codes still accepted for historical compatibility.
        budget:
          type: string
          minLength: 1
          maxLength: 280
          description: Budget band label (Under $5K, $5,000–$15,000, …) or legacy exact amount.
        budgetBand:
          type: string
          enum: [UNDER_5K, BETWEEN_5K_15K, BETWEEN_15K_50K, OVER_50K]
          description: Optional structured band; normalized from budget label when omitted.
        budgetCents:
          type: integer
          minimum: 1
          description: Exact budget in integer cents (legacy compatibility).
        attachments:
          type: array
          maxItems: 5
          items:
            type: object
            required: [downloadUrl, fileName]
            properties:
              downloadUrl:
                type: string
                format: uri
              fileName:
                type: string
              mimeType:
                type: string
              sizeBytes:
                type: integer
                maximum: 20971520
        timeline:
          type: string
          description: Structured timeline label (ASAP, Within 2 weeks, …) or legacy YYYY-MM-DD date.
        urgency:
          type: string
          minLength: 1
          maxLength: 280
          description: Emergency, High, Medium, or Low (canonical codes also accepted).
        description:
          type: string
          minLength: 10
          maxLength: 4000
        externalContractorId:
          type: string
          minLength: 1
          maxLength: 160
      allOf:
        - if:
            properties:
              source:
                const: direct-contractor-profile-request
          then:
            required:
              - externalContractorId
        - if:
            properties:
              source:
                const: general/get-three-estimates
          then:
            not:
              required:
                - externalContractorId
    EstimateAccepted:
      type: object
      additionalProperties: false
      required:
        - ok
        - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: false
          required:
            - leadId
            - replay
            - reviewStatus
            - blockers
          properties:
            leadId:
              type: string
            replay:
              type: boolean
            reviewStatus:
              type: string
              enum: [pending_review, routed]
            blockers:
              type: array
              items:
                type: string
                enum:
                  - budget_review
                  - pricing_review
                  - category_review
                  - contractor_review
                  - attachment_ingestion
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - ok
        - error
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - integration_disabled
                - unauthorized
                - unsupported_media_type
                - invalid_json
                - validation_error
                - invalid_reference
                - idempotency_conflict
                - internal_error
            message:
              type: string
            issues:
              type: array
              items:
                type: object
                required:
                  - path
                  - message
                properties:
                  path:
                    type: string
                  message:
                    type: string